src/main/java/ru/yandex/practicum/filmorate/dao/mpa/MpaDaoImpl.java
package ru.yandex.practicum.filmorate.dao.mpa;

import lombok.RequiredArgsConstructor;
import org.springframework.jdbc.core.JdbcTemplate;
import org.springframework.jdbc.core.RowMapper;
import org.springframework.stereotype.Component;
import ru.yandex.practicum.filmorate.model.Mpa;

import java.util.List;

@Component
@RequiredArgsConstructor
public class MpaDaoImpl implements MpaDao {

    private final JdbcTemplate jdbcTemplate;

    @Override
    public List<Mpa> getAllMpa() {
        return jdbcTemplate.query("select * from mpa_ratings", mpaRowMapper());
    }

    @Override
    public Mpa getMpaById(Long id) {
        return jdbcTemplate.queryForObject("select * from mpa_ratings where id = ?", mpaRowMapper(), id);
    }

    private RowMapper<Mpa> mpaRowMapper() {
        return (rs, rowNum) -> new Mpa(
                rs.getLong("id"),
                rs.getString("name")
        );
    }
}